# Optimized Quaternion to Matrix3/Matrix4 conversions (~80% faster).