--- a/OpenTK.Benchmarks/OpenTK.Benchmarks.csproj
+++ b/OpenTK.Benchmarks/OpenTK.Benchmarks.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\src\OpenTK.Mathematics\OpenTK.Mathematics.csproj" />
+  </ItemGroup>
+
+</Project>