--- a/OpenTK.Benchmarks/Program.cs
+++ b/OpenTK.Benchmarks/Program.cs
@@ -1,0 +1,15 @@
+﻿using BenchmarkDotNet.Running;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OpenTK.Benchmarks
+{
+    public class Program
+    {
+        public static void Main()
+        {
+            BenchmarkRunner.Run<MatrixBenchmark>();
+        }
+    }
+}