--- a/src/OpenTK.Mathematics/Matrix/Matrix3.cs
+++ b/src/OpenTK.Mathematics/Matrix/Matrix3.cs
@@ -514,8 +514,35 @@ namespace OpenTK.Mathematics
         /// <param name="result">Matrix result.</param>
         public static void CreateFromQuaternion(in Quaternion q, out Matrix3 result)
         {
-            q.ToAxisAngle(out Vector3 axis, out float angle);
-            CreateFromAxisAngle(axis, angle, out result);
+            // Adapted from https://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation#Quaternion-derived_rotation_matrix
+            float sqx = q.X * q.X;
+            float sqy = q.Y * q.Y;
+            float sqz = q.Z * q.Z;
+            float sqw = q.W * q.W;
+
+            float xy = q.X * q.Y;
+            float xz = q.X * q.Z;
+            float xw = q.X * q.W;
+
+            float yz = q.Y * q.Z;
+            float yw = q.Y * q.W;
+
+            float zw = q.Z * q.W;
+
+            float s2 = 2f / (sqx + sqy + sqz + sqw);
+
+            result.Row0.X = 1f - (s2 * (sqy + sqz));
+            result.Row1.Y = 1f - (s2 * (sqx + sqz));
+            result.Row2.Z = 1f - (s2 * (sqx + sqy));
+
+            result.Row0.Y = s2 * (xy - zw);
+            result.Row1.X = s2 * (xy + zw);
+
+            result.Row2.X = s2 * (xz - yw);
+            result.Row0.Z = s2 * (xz + yw);
+
+            result.Row2.Y = s2 * (yz + xw);
+            result.Row1.Z = s2 * (yz - xw);
         }
 
         /// <summary>