### Purpose of this PR

Added tests for these conversions to make them closely match the previous functions values (the new function has greater accuracy with axis aligned rotations).
Added benchmark project with performance benchmarks for these functions.

### Testing status

Added tests for this, they pass.
Also added performance benchmarks to show this code is faster.
The following is from a benchmark run for the Quaternion to Matrix4 benchmarks:
```
|                   Method |                 Quat |     Mean |    Error |   StdDev | Ratio |
|------------------------- |--------------------- |---------:|---------:|---------:|------:|
|    Matrix4FromQuaternion | V: (9(...)2E+09 [63] | 22.25 ns | 0.479 ns | 0.551 ns |  0.24 |
| Matrix4FromQuaternionOld | V: (9(...)2E+09 [63] | 91.55 ns | 1.754 ns | 1.877 ns |  1.00 |
```