@NogginBops 2020-11-08
Fixed a license issue where the original code was adapted from GPL licensed code. This new version is adapted from [this](https://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation#Quaternion-derived_rotation_matrix) wikipedia page.