@NogginBops 2020-11-12
It seems like there might have been an issue with this pr that was not caught by the unit tests or my manual testing.
I'm investigating what this issue might be and will make a follow up pr/comment.